23.06.22

+++ server/src/main/java/seb44pre036/qna/answer/service/AnswerService.java
@@ -11,6 +11,8 @@ import seb44pre036.qna.member.service.MemberService;
 import seb44pre036.qna.question.entity.Question;
 import seb44pre036.qna.question.service.QuestionService;
 
+import java.io.IOException;
+import java.time.LocalDateTime;
 import java.util.Optional;
 
 @Service
@@ -20,34 +22,45 @@ public class AnswerService {
     private QuestionService questionService;
 
     @Autowired
-    public void AnswerService(AnswerRepository answerRepository, MemberService memberService /*,QuestionService questionService*/){
+    public void AnswerService(AnswerRepository answerRepository, MemberService memberService, QuestionService questionService) {
         this.answerRepository = answerRepository;
         this.memberService = memberService;
         this.questionService = questionService;
     }
 
-    public Answer postAnswer (Answer answer){
+    public Answer postAnswer(Answer answer) {
+        // 질문 등록 여부 검사
+        Question question = questionService.findVerifiedQuestion(answer.
+                getQuestion().
+                getQuestionId());
+
+        // 답변 등록자
+        Member user = memberService.findVerifiedMember(answer.
+                getMember().
+                getMemberId());
+
+        // 질문 작성자
+        Member writer = memberService.
+                findVerifiedMember(question.getMember().getMemberId());
 
         // 본인 질문에 답변 등록 불가
-        if(answer.getMember().getMemberId() == answer.getQuestion().getMember().getMemberId()){
+        if (writer.getMemberId() == user.getMemberId()) {
             throw new BusinessLogicException(ExceptionCode.ANSWER_NOT_CREATED);
+        } else {
+            Answer createAnswer = answerRepository.save(answer);
+            // 회원이 작성한 답변 추가
+            user.addAnswer(createAnswer);
+            // 질문에 작성된 답변 추가
+            //question.addAnswer(createAnswer);
+            return createAnswer;
         }
-
-        Answer createAnswer = answerRepository.save(answer);
-
-        // 맴버에 답변 추가
-        // answer.getMember().addAnswer(createAnswer);
-
-
-        return createAnswer;
     }
 
-    public Answer findAnswer(long answerId){
-
+    public Answer findAnswer(long answerId) {
         return findVerifiedAnswer(answerId);
     }
 
-    public Answer findVerifiedAnswer(long answerId){
+    public Answer findVerifiedAnswer(long answerId) {
         Optional<Answer> optionalAnswer = answerRepository.findById(answerId);
         Answer findAnswer =
                 optionalAnswer
@@ -57,33 +70,50 @@ public class AnswerService {
         return findAnswer;
     }
 
-    public Answer updateAnswer(Answer answer){
-        // 답변자일 경우 수정 가능
-
-        // 존재하는 답변인지 확인
+    public Answer updateAnswer(long memberId, Answer answer) {
         Answer preAnswer = findVerifiedAnswer(answer.getAnswerId());
 
-        // update
-        Optional.ofNullable(answer.getContent()).ifPresent(newContent -> preAnswer.setContent(newContent));
-        Optional.ofNullable(answer.getUpdatedAt()).ifPresent(updateAt -> preAnswer.setUpdatedAt(updateAt));
+        // 답변 최초 작성자
+        final long writerId = preAnswer.getMember().getMemberId();
 
-        return answerRepository.save(preAnswer);
+        // 답변자 본인일 경우 수정 가능
+        if (memberId == writerId) {
+            // update
+            Optional.ofNullable(answer.getContent()).ifPresent(newContent -> preAnswer.setContent(newContent));
+            Optional.ofNullable(LocalDateTime.now()).ifPresent(updateAt -> preAnswer.setUpdatedAt(updateAt));
+            return answerRepository.save(preAnswer);
+        } else {
+            throw (new BusinessLogicException(ExceptionCode.ACCESS_DENIED));
+        }
     }
 
-    public void deleteAnswer(long answerId){
-        // 질문 작성자, 답변자일 경우 삭제 가능
-
+    public void deleteAnswer(long answerId,long memberId) {
         Answer answer = findAnswer(answerId);
-        answerRepository.delete(answer);
+        long writerId = answer.getMember().getMemberId();
+
+        // 답변자일 경우 삭제 가능
+        if(memberId == writerId){
+            answerRepository.delete(answer);
+        }
+        else{
+            throw(new BusinessLogicException(ExceptionCode.ACCESS_DENIED));
+        }
     }
 
-    public Answer selectAnswer(long answerId){
-        // 본인 답변에 채택 불가
+    public Answer selectingAnswer(long answerId, long memberId) {
 
         //존재하는 답변인지 확인
         Answer answer = findAnswer(answerId);
-        //Optional.of(Answer.AnswerStatus.ANSWER_SELECT).ifPresent((select) -> answer.setAnswerStatus(select));
 
-        return answer;
+        // 질문 작성자
+        final long writerId = answer.getQuestion().getMember().getMemberId();
+
+        // 질문자일 경우에만 채택 가능
+        if (memberId != writerId) {
+            throw (new BusinessLogicException(ExceptionCode.ACCESS_DENIED));
+        } else {
+            answer.setAnswerStatus(Answer.AnswerStatus.ANSWER_SELECT);
+            return answerRepository.save(answer);
+        }
     }
 }
